+++ test/test_helper.exs
@@ -81,6 +81,8 @@ ExUnit.start(
     # :individual_identity_log,
     # :business_identity,
     # :business_identity_log,
+    # :business_attachment,
+    # :business_attachment_log,
   ]
 )
 
@@ -106,3 +108,4 @@ Code.require_file("./test/utils/individual_account_request.exs")
 Code.require_file("./test/utils/individual_account_attachment.exs")
 Code.require_file("./test/utils/individual_identity.exs")
 Code.require_file("./test/utils/business_identity.exs")
+Code.require_file("./test/utils/business_attachment.exs")