+++ CHANGELOG.md
@@ -84,6 +84,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IndividualIdentity.Log sub-resource
 - BusinessIdentity resource
 - BusinessIdentity.Log sub-resource
+- BusinessAttachment resource
+- BusinessAttachment.Log sub-resource
 ### Changed
 - IssuingBin resource to IssuingProduct
 - settlement parameter to funding_type and client parameter to holder_type of IssuingProduct